Fix service-guid counting bug
Initial encounter of service-guid should be set to 1, not 0

+++ solution/src/EventFilter.java
@@ -56,7 +56,7 @@ public class EventFilter {
 				count++;
 				serviceGuids.put(serviceGuid, new Integer(count));
 			} else {
-				serviceGuids.put(serviceGuid, new Integer(0));
+				serviceGuids.put(serviceGuid, new Integer(1));
 			}
 		}
